TSF-PyTorch-Pipeline/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_py-xgboost-mutex=2.0=cpu_0
adtk=0.6.2=py_0
argon2-cffi=21.3.0=pyhd8ed1ab_0
argon2-cffi-bindings=21.2.0=py38h294d835_2
asttokens=2.0.5=pyhd8ed1ab_0
attrs=21.4.0=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
beautifulsoup4=4.11.1=pyha770c72_0
blas=1.0=mkl
bleach=5.0.0=pyhd8ed1ab_0
brotli=1.0.9=h8ffe710_7
brotli-bin=1.0.9=h8ffe710_7
brotlipy=0.7.0=py38h294d835_1004
ca-certificates=2022.5.18.1=h5b45459_0
certifi=2022.5.18.1=py38haa244fe_0
cffi=1.15.0=py38hd8c33c5_0
charset-normalizer=2.0.12=pyhd8ed1ab_0
colorama=0.4.4=pyh9f0ad1d_0
cryptography=37.0.2=py38hb7941b4_0
cudatoolkit=11.3.1=h59b6b97_2
cycler=0.11.0=pyhd8ed1ab_0
debugpy=1.6.0=py38h885f38d_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
einops=0.4.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_0
et_xmlfile=1.0.1=py_1001
executing=0.8.3=pyhd8ed1ab_0
flit-core=3.7.1=pyhd8ed1ab_0
fonttools=4.33.3=py38h294d835_0
freetype=2.10.4=h546665d_1
gettext=0.19.8.1=ha2e2712_1008
gst-plugins-base=1.20.2=he07aa86_1
gstreamer=1.20.2=hdff456e_1
icu=70.1=h0e60522_0
idna=3.3=pyhd8ed1ab_0
importlib-metadata=4.11.4=py38haa244fe_0
importlib_resources=5.7.1=pyhd8ed1ab_1
intel-openmp=2022.1.0=h57928b3_3787
ipykernel=6.13.0=py38h4317176_0
ipython=8.4.0=py38haa244fe_0
ipython_genutils=0.2.0=py_1
ipywidgets=7.7.0=pyhd8ed1ab_0
jedi=0.18.1=py38haa244fe_1
jinja2=3.1.2=pyhd8ed1ab_1
joblib=1.1.0=pyhd8ed1ab_0
jpeg=9e=h8ffe710_1
jsonschema=4.6.0=pyhd8ed1ab_0
jupyter=1.0.0=py38haa244fe_7
jupyter_client=7.3.1=pyhd8ed1ab_0
jupyter_console=6.4.3=pyhd8ed1ab_0
jupyter_core=4.10.0=py38haa244fe_0
jupyterlab_pygments=0.2.2=pyhd8ed1ab_0
jupyterlab_widgets=1.1.0=pyhd8ed1ab_0
kiwisolver=1.4.2=py38hbd9d945_1
krb5=1.19.3=h1176d77_0
lcms2=2.12=h2a16943_0
lerc=3.0=h0e60522_0
libblas=3.9.0=14_win64_mkl
libbrotlicommon=1.0.9=h8ffe710_7
libbrotlidec=1.0.9=h8ffe710_7
libbrotlienc=1.0.9=h8ffe710_7
libcblas=3.9.0=14_win64_mkl
libclang=13.0.1=default_h81446c8_0
libdeflate=1.10=h8ffe710_0
libffi=3.4.2=h8ffe710_5
libglib=2.70.2=h3be07f2_4
libiconv=1.16=he774522_0
liblapack=3.9.0=14_win64_mkl
libogg=1.3.4=h8ffe710_1
libpng=1.6.37=h1d00b33_2
libsodium=1.0.18=h8d14728_1
libtiff=4.3.0=hc4061b1_4
libuv=1.43.0=h8ffe710_0
libvorbis=1.3.7=h0e60522_0
libwebp=1.2.2=h57928b3_0
libwebp-base=1.2.2=h8ffe710_1
libxcb=1.13=hcd874cb_1004
libxgboost=1.5.1=cpu_h110a20b_2
libzlib=1.2.12=h8ffe710_0
local-attention=1.4.3=pypi_0
lz4-c=1.9.3=h8ffe710_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markupsafe=2.1.1=py38h294d835_1
matplotlib=3.5.2=py38haa244fe_0
matplotlib-base=3.5.2=py38he529843_0
matplotlib-inline=0.1.3=pyhd8ed1ab_0
mistune=0.8.4=py38h294d835_1005
mkl=2022.0.0=h0e2418a_796
msys2-conda-epoch=20160418=1
munkres=1.1.4=pyh9f0ad1d_0
nbclient=0.6.4=pyhd8ed1ab_1
nbconvert=6.5.0=pyhd8ed1ab_0
nbconvert-core=6.5.0=pyhd8ed1ab_0
nbconvert-pandoc=6.5.0=pyhd8ed1ab_0
nbformat=5.4.0=pyhd8ed1ab_0
nest-asyncio=1.5.5=pyhd8ed1ab_0
notebook=6.4.11=pyha770c72_0
numpy=1.22.4=py38h1d2777f_0
openjpeg=2.4.0=hb211442_1
openpyxl=3.0.9=pyhd8ed1ab_0
openssl=1.1.1o=h8ffe710_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.4.2=py38hcc40339_2
pandoc=2.18=h57928b3_0
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.3=pyhd8ed1ab_0
patsy=0.5.2=pyhd8ed1ab_0
pcre=8.45=h0e60522_0
pickleshare=0.7.5=py_1003
pillow=9.1.1=py38hd8e0db4_0
pip=21.2.2=py38haa95532_0
plotly=5.8.0=pyhd8ed1ab_1
prometheus_client=0.14.1=pyhd8ed1ab_0
prompt-toolkit=3.0.29=pyha770c72_0
prompt_toolkit=3.0.29=hd8ed1ab_0
psutil=5.9.1=py38h294d835_0
pthread-stubs=0.4=hcd874cb_1001
pure_eval=0.2.2=pyhd8ed1ab_0
py-xgboost=1.5.1=cpu_py38hc22ac66_2
pycparser=2.21=pyhd8ed1ab_0
pygments=2.12.0=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.9=pyhd8ed1ab_0
pyqt=5.15.4=py38h75e37d8_1
pyqt5-sip=12.9.0=py38h885f38d_1
pyrsistent=0.18.1=py38h294d835_1
pysocks=1.7.1=py38haa244fe_5
python=3.8.0=hff0d562_2
python-dateutil=2.8.2=pyhd8ed1ab_0
python-fastjsonschema=2.15.3=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytorch=1.11.0=py3.8_cuda11.3_cudnn8_0
pytorch-mutex=1.0=cuda
pytz=2022.1=pyhd8ed1ab_0
pywin32=303=py38h294d835_0
pywinpty=2.0.2=py38h5da7b33_0
pyzmq=23.1.0=py38h09162b1_0
qt-main=5.15.3=h467ea89_0
qtconsole=5.3.0=pyhd8ed1ab_0
qtconsole-base=5.3.0=pyha770c72_0
qtpy=2.1.0=pyhd8ed1ab_0
requests=2.27.1=pyhd8ed1ab_0
scikit-learn=1.1.1=py38h9235413_0
scipy=1.8.1=py38h9bf8e03_0
seaborn=0.11.2=hd8ed1ab_0
seaborn-base=0.11.2=pyhd8ed1ab_0
send2trash=1.8.0=pyhd8ed1ab_0
setuptools=61.2.0=py38haa95532_0
sip=6.5.1=py38h885f38d_2
six=1.16.0=pyh6c4a22f_0
soupsieve=2.3.1=pyhd8ed1ab_0
sqlite=3.38.3=h2bbff1b_0
stack_data=0.2.0=pyhd8ed1ab_0
statsmodels=0.10.2=py38hc8d92b1_0
tabulate=0.8.9=pyhd8ed1ab_0
tbb=2021.5.0=h2d74725_1
tenacity=8.0.1=pyhd8ed1ab_0
terminado=0.15.0=py38haa244fe_0
threadpoolctl=3.1.0=pyh8a188c0_0
tinycss2=1.1.1=pyhd8ed1ab_0
tk=8.6.12=h8ffe710_0
toml=0.10.2=pyhd8ed1ab_0
torchaudio=0.11.0=py38_cu113
torchsummary=1.5.1=pypi_0
torchvision=0.12.0=py38_cu113
tornado=6.1=py38h294d835_3
tqdm=4.64.0=pyhd8ed1ab_0
traitlets=5.2.2.post1=pyhd8ed1ab_0
typing_extensions=4.2.0=pyha770c72_1
unicodedata2=14.0.0=py38h294d835_1
urllib3=1.26.9=pyhd8ed1ab_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.6.0=py38haa244fe_0
win_inet_pton=1.1.0=py38haa244fe_4
wincertstore=0.2=py38haa95532_2
winpty=0.4.3=4
xgboost=1.5.1=cpu_py38hc22ac66_2
xlrd=2.0.1=pyhd8ed1ab_3
xorg-libxau=1.0.9=hcd874cb_0
xorg-libxdmcp=1.1.3=hcd874cb_0
xz=5.2.5=h62dcd97_1
zeromq=4.3.4=h0e60522_1
zipp=3.8.0=pyhd8ed1ab_0
zlib=1.2.12=h8ffe710_0
zstd=1.5.2=h6255e5f_1
